pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cortexia"
version = "1.0.0"
description = "Face recognition system with trust scoring and anti-spoofing"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Raghav Shukla"},
]
keywords = ["face-recognition", "anti-spoofing", "computer-vision", "deep-learning", "biometrics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.2.5",
    "redis>=5.0.0",
    "celery[redis]>=5.3.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.9.0",
    "insightface>=0.7.3",
    "onnxruntime>=1.17.0",
    "scikit-learn>=1.4.0",
    "hdbscan>=0.8.33",
    "Pillow>=10.2.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "structlog>=24.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "click>=8.1.0",
    "rich>=13.7.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "black>=24.1.0",
    "pre-commit>=3.6.0",
    "locust>=2.23.0",
    "factory-boy>=3.3.0",
]
gpu = [
    "onnxruntime-gpu>=1.17.0",
]

[project.scripts]
cortexia = "cortexia.cli:cli"

[tool.setuptools.packages.find]
include = ["cortexia*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short --cov=cortexia --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py311"]
